Do not crash when looking at the hierarchy of a tracker with cross-project children
Instead of crashing the tracker administration interface just hide them.
This is not great but it should be good enough for now. Doing more means
that we need a way to distinguish trackers from another projects in this
section and also that we need to take into account that a user might not
be an admin of the cross-project tracker.

Part of story #17882: define cross project hierarchy between Features and User Stories

Change-Id: I131171f12943e3f0825a067c958aaa74dd8006f0