Fix hard crash when running the unit tests
The test suite can now be executed completly under PHP 8 (with errors)
without a hard crash in the middle.

The crash was caused by code generated by Mockery to create the spy of
the ZipArchive. There is no value of having this archive as a mock/spy,
the real thing can be used directly.

Part of request #17931: Unit tests should pass with PHP 8.0

Change-Id: Ib4f0cc84948934cc28010fa57ed1ce65d17e1fd4