Redirect to milestone after backlog item creation
Go to a pane other than pv2 (like overview or cardwall) on a given
milestone
Use the +New dropdown menu to add a new backlog item.
=> since there is no modal on this pane, you are redirected to the
tracker in order to create the artifact.
Fill the form and click submit
=> once the artifact created, you are redirected to the previous pane,
so the chance to loose the context is decreased.

Note: the behavior is kept if the user choose "submit and continue" to
add multiple backlog items to a milestone.

(Other panes like taskboard or testplan will be handled in a separate
commit to not clutter the review)

For the record, we cannot use the already existing mechanism
"link-artifact-id" (in combination with the peculiar "immediate"
parameter) since it is used for legacy overlay and does not manage well
the display in case of form error at artifact submission. Therefore we
have to introduce an AD specific parameter.

Part of story #16206: « + New » button, better contextualisation

Change-Id: Id711280f05308a18c03c82575af793f53641632c