request #18400: Convert mailing list to gettext
No functional changes.

The only entry that was marked as overridable was the mention
about Mailman and Python (only in english, empty in french):

> Mailing lists provided via a GNU Mailman. Thanks to the Mailman and
> Python crews for excellent software.

Those thanks does not bring any value to the end user in the interface
(and GNU Mailman is already mentionned in the documentation). Therefore
I decided to not display it anymore if it has not been customized.

Note: some entries in mail.tab where not related to mailing list, but I
converted them too. If it complexifies too much the review, I can split
the commit.

Change-Id: I61a06936d07ed5f8312f69f8c59431a3bb659a4f