request #18404: Update to Psalm 4.3.1
Release notes:
https://github.com/vimeo/psalm/releases/tag/4.3.0
https://github.com/vimeo/psalm/releases/tag/4.3.1

Most of the new errors are caused by the fact that RedundantCast is now
a separate issue (it's no more bundled with RedundantCondition).
The dockblock of Group::getID has been updated to include the
possibility of returning a string, the existing docblock, while correct
in a few cases, was saying only a int is returned. This was generating a
lot of incorrect RedundantCast issues.

Change-Id: I48c80659b0a2e24c24abd230587a5d4c9e947ccc