Remove direct inclusion of jpgraph.php in class files
The direct inclusions have been added in
1b14ec815dd92eb78972e4be8d7a8c7d307dccd8 to solve Psalm issues that are
caused by the convoluted way JPGraph is loading its symbols.
This contribution rollbacks to the old way and just ignore the errors.
Directly including jpgraph.php from class files is not fine because it
redefines the error handler which can cause unit tests to pass when they
should be failing. Psalm is right about the issues and both situations
are not great so let's use the less annoying for now.

Part of request #17148: Tracker move to PSR4

Change-Id: I5087f38559a926e30475ea1eaa8501db4e557dc8