Information message when coming from milestone
When using the +New dropdown button to create an backlog item in a
milestone, for example in the overview pane, then a message informs the
user and contains a link to the milestone for which we are creating a
new backlog item.

This is the same mechanism that has been used for test plan (to create
tests for backlog items). See git #tuleap/stable/579877d93190b73c74762c6c3ec7415ec58f14e2 for example[0]

Part of story #16206: « + New » button, better contextualisation

[0] https://gerrit.tuleap.net/c/tuleap/+/19658

Change-Id: I1732d7de7cfe03637ac3acb6a31190637a008570