Generate the "OIDC authentication state" only when the user explicitely asks to be authenticated
Currently Tuleap generates this state every time the URLs to start the
authentication flow needs to be displayed. This can be problematic when
another request is done after the one used to display the page initially
(e.g. to display a user avatar). With this contribution, the user needs
to start the authentication flow to overwrite the "OIDC authentication
state".

To reproduce the issue:
 1. without being authenticated open the homepage in tab A
 2. without being authenticated open the homepage in tab B
 3. go back to tab A, do not refresh, try to authenticate against an
    OIDC provider

request #18410: Open ID Connect Client+ Unique + Anonymous homepage + News => impossible to login

Change-Id: Ifcc0de35af5fe3cda6dafb6c76f68d1e3aac069c