Fix fatal error when reaching an invalid milestone
Instead of building a pane for a non existing artifact, we should
display an error instead.

To test, go to the details pane of a valid milestone, look at the url,
and change aid= parameter value to an unexisting artifact, or -1.
=> you are redirected to the top planning of the project, with an error
message.

You can also hack the redirection url that is used to link a backlog
item to the milestone, for example from the details pane:

  planning[details][1]=:fake_id_milestone&link-to-milestone=1

Note: this allows us to remove the last bit of dead code that was
removed in git #tuleap/stable/54e1d75df04c23057343679ff102f8c8e5939723 [0]

Part of story #16206: « + New » button, better contextualisation

[0] https://gerrit.tuleap.net/c/tuleap/+/20968

Change-Id: I991d4575927546d264a791c062a92a664902de83