Introduce an end-to-end test for the creation artifact modal
Part of request #19273 Have an end-to-end test for the creation artifact
modal

It does not look like this matches a test_case, I'll create one.

How to test:
- "make tests_cypress" should succeed

Notes:
The "file" field is not tested yet. It looks like it may be possible to
test file upload through cypress but I would prefer to try this in a
dedicated patch as this test is already pretty huge.

Please also note that I have purposefully "cheated" for ListPicker and
date picker by using "force: true". The purpose of this test is not to
test in detail how the date picker or list picker works. We should write
integration tests dedicated to these components. I've made an exception
for Select2 in open lists because there is no way to "cheat" for open
list bound to ugroups. It fetches users by autocompletion, so I cannot
use .select().

Change-Id: Ifcb3b12173504b1c985c119641bea04609ff7456