Redirect to taskboard after card edition
Since there isn't any modal to edit card on taskboard yet, we redirect users
to tracker service in order to edit a card (story, task, …). Instead of
simply staying in tracker, this contribution redirects the user to the
taskboard after the artifact has been updated.

This is the same behavior than the overview pane.

Part of story #16206: « + New » button, better contextualisation

Change-Id: I241501424906bfc79a306cca3e4c4da3252eae0a