Fix issues detected by Psalm in the SourceArtifactNatureAnalyzer class
The class was introduced on an older version of Psalm which detected
those issues under another rule name.

The docblock of the Tracker::getId class was fixed to match the reality
and pre-existing errors have been integrated into the baseline.
Took the opportunity to also fix PFUser::getId which is another big
offender of the same issue.

Part of story #18405: distinguish Features and P.I. that comes from a Program in a Team

Change-Id: I525d1a7e05184eea76b8fdec92c3353991e23599