Extract ckeditor image upload lib
Part of story #18336 add support of markdown

How to test:
- No functional change in "Image drag/drop" feature in the Artifact
Modal, TestManagement or the Tracker Artifact view

Notes: Since we are going to change the CKEditors in Trackers to add a
new Markdown format, extracting this lib will get it out of our way.
CKEditor with Markdown is likely to require a lib itself, dependencies
will be clearer that way.

As a bonus, it will also help us for request #18391 Migrate Tracker
Artifact Modal out of AngularJS. The Artifact modal depends on this lib.

regenerator/runtime is required to just import the files in Jest
(otherwise there's an error), so whenever we need the Error classes and
cannot mock the module completely, we had to add a dependency :/ that is
to say: in core (for ckeditor) and in Trackers for the Artifact Modal.

Change-Id: I5639d0f2a45a2fe21a4582fbea1d81826e3c12b8