Introduce Administration page for Mailing Lists
Current mailing lists administration is done through three different
pages.
1. A page to create mailing lists
2. A page to edit all mailing lists (but only one ML at a time)
3. An intermediate page that links to the previous ones. This one is not
accessible unless you know the url. The only usage is in the e2e tests.

This way of administrate a service does not match our latest standards.

The proposal is to have an administration page that lists the existing
mailing lists and buttons to add or edit mailing list, the TLP way.

This commits focuses only on the introduction of this new page. Mailing
lists are listed in a basic table, and a button links to the
"add lists" page that has recently been converted to burning parrot[0]

In a next contribution, we will enhance the table by adding links to the
list and rework the way the list are updated.

Part of request #18425: Mailing lists should use BurningParrot

[0] git #tuleap/stable/9dcfe5abc487bff31a2deb4a7426c3b15a97ba90
https://gerrit.tuleap.net/c/tuleap/+/21018

Change-Id: I579c766e0a89742b929f17d3c5ef46fb90291acc