request #18441: Increase password work factor
The cost of Bcrypt hashed passwords has been increased to 13.
This follow the current recommendations [0] and the example of other
major web frameworks [1].

The existing passwords will be rehashed when the users log in.

The work factor for password stored for the UNIX accounts has been
increased by ~ a similar factor.

[0] https://cheatsheetseries.owasp.org/cheatsheets/Password_Storage_Cheat_Sheet.html#bcrypt
[1] https://github.com/symfony/symfony/blob/8b51547061b29081803eb9fdf9b02cf80a1e1a74/src/Symfony/Component/Security/Core/Encoder/NativePasswordEncoder.php#L35

Change-Id: I0bdc4f9cdaa262915ea14f43beccd73f2a6176f8