Update of SVN apache config should not depend on CVS or home page
Until now most failures during the execution of SYSTEM_CHECK were
fail fast => the SYSTEM_CHECK was aborted with the first error found.

This is no longer the desired behaviour as a bunch of things are done
in plugins that are called at the end of the process. Some may even
argue it was not really an intentional behaviour in first place.

Some behaviours might change with this patch on weirdly configured
platform. Until now they might have only one error in SYSTEM_CHECK
=> they will now see them all.

Note: this patch comes with bunch of changes related to CVS that were
triggered by psalm due to code move. If it's too much for a patch,
I'll split in two.

request #18430 TEE Docker Image - Problem on existing SVN repository

Change-Id: I1466e0857d701988f1200a650d760f8fec3a3e3a