Custom session should not rely on a PHP warning
This warning is now replaced by an explicit exception.
With PHP 8.0 this warning is now a \TypeError so it cannot be considered
as inconsequential anymore [0].

It does not look like the warning is triggered on production, at least
it does not appears in the logs of Tuleap servers Enalean manages
directly.

In any case it should be considered as a bug since it indicates somethin
unexpected happen in the code.

The test should now pass with PHP 8.0:
    make tests-unit-php PHP_VERSION=80 FILES=tests/unit/common/session/Codendi_SessionTest.php

Part of request #17931: Unit tests should pass with PHP 8.0

[0] https://github.com/php/php-src/blob/php-8.0.0/UPGRADING#L113-L114

Change-Id: Ifcab130f6613d7a7f756309f0e7f8db18213382a