request #18456 Webdav plugin should expose URLs compliant with WebDav protocol
Replace "direct" `/plugins/webdav` access with nginx configuration by a front
route that allows to deal with WebDAV protocol (ending `/` is not allowed).

The URLVerification is simplified to only deal with URLs that are addressed
to a webdav host different of main host.

The configuration for webdav is simplified.

On main domain (says tuleap.example.com)
the `/plugins/webdav` URL is always available and is the recommanded way
to access webdav share.

For backward compatibility reasons, access to dedicated domain
(says webdav.tuleap.example.com) is kept and is only possible on
`/` path.

Hence:
- `$webdav_base_uri` is no longer configurable (and removed from default conf)
- `$webdav_host` is removed from conf file example to avoid new usage

Change-Id: Ie991f06844fb5fb5331ab50aff1bf1507b9e1da7