Group commit references by section
Functional changes
==================

Make sure that an artifact is referenced by (or is referencing) two or
more git commits, in different repositories. On the cross reference
field, the list of commits are grouped by sections with the title of
the section being the path of the repository.

Implementation details
======================

Previous implementation of cross references display in trackers was
based on the legacy way that merely didn't change for a decade. Now that
we want to display cross references in a structured manner, our objects
naming must map for what they are used. So the proposal here is to have
the following architecture:

```
      +--------+
      | Nature |
      |--------|
      | Label  |
      | Icon   |
      +--------+
          |
          |
          | n
     +---------+
     | Section |
     |---------|
     | Label   |
     +---------+
          |
          |
          | n
  +----------------+
  | CrossReference |
  |----------------|
  | title          |
  | url            |
  | delete_url     |
  +----------------+
```

…knowing that we are starting from a flat array of cross references. So
we need to ask to all "owners" to organize their own cross references
into sections/nature. For now, only git plugin is involved, but in next
commits tracker, pullrequest will join the party.

In order to not be biaised by the existing goofy structure brought by
CrossReferenceFactory (which are doing sql calls directly, and are
providing unwanted semi-organized cross references collection), we are
starting from the ground up.

Therefore this reorganization produces a huge changeset. I tried to
split the commit, but didn't find a way to produce something small AND
understandable without losing my mind :/ For my defense, it should be
noted that nearly two thirds of the added lines in this changeset are unit
tests :)

---

Part of story #17180 take into account gitlab commits

Change-Id: Ibdab459f5a3ca5e8e22a427687ff2f2ad4c74d4b