request #17996 [Kanban] Open the "backlog" column by default
When we add a new element in the Kanban with the "+New" button that directly
goes to the Backlog, given that this colum is collapsed by default, users don't see their
newly created item.

We change this by opening the Backlog column by default in Kanban.
User preference is still taken into account if defined.

Change-Id: I620e296a0699a877be187cab5168db4ee3fdb7b6