Pimp my artifact cross reference
When there is a cross reference between two artifacts, this cross
reference now displays a title badge (the xref) and the artifact title
instead of just the legacy xref value.

If user cannot access to the artifact, then the reference is not
displayed anymore.

If user cannot access to the title of the artifact (no perms for the
field), then the title is empty, but the badge is still displayed.

Part of story #17180 take into account gitlab commits

Change-Id: I0983a2fbd7d81edc8f96459432968f68f4857cf4