Webdav cannot be accessed by anonymous
Unlike advertise by the comments, it's never been possible to access
webdav as anonymous because the username was always required.

Commit 7b166353f057dc70c aligned the code behaviour to the comment but
it's unclear from a functional stand-point if it ever was wanted and
it's safer today to revert to the previous behaviour: no anonymous access.

The corresponding code was also removed from WebDAVRoot to avoid
confusions.

Part of story #18435: access webdav with a personal access key

Change-Id: I4ef6c964592dd8397265b790646e8ed895faa761