Refactor WebdavRoot and WebdavProject to make dependency explicit
In order to make tests more effective (less partial mocks) and
to better grasp the webdav internals vs. sabre/dav contract the
first 2 nodes of the graph are refactored to have most of the
dependencies explicit and make everything that can be private
private.

Part of request #18470 Bump sabre/dav to 4.1.3 🙄

Change-Id: Id6b7b2abf4ab699842528a8ecfb6156e0ecbf3b9