Display commit information instead of raw cross ref label
For each commit that is referencing (or is being referenced by) an
artifact, in addition of the sha1 as badge, we now display:
* The commit title instead of the reference
* The author and the date

Note:
Integrating tlp-avatar is more complicated than integrating tlp badges
or panes. Therefore, for now the tlp-avatar-mini rules have been
copy-pasted. Maybe the situation will be better when we will integrate
the new cross references styling in core…

Part of story #17180 take into account gitlab commits

Change-Id: I49ea9c3253356be929c18d09d939408218b7bbf5