Add project background information in the milestone and backlog item representations
To have the capacity of changing the color of the element when we are
dealing with elements related to another project we need to have
information about the project background color. This information is now
available in the REST representations.

Part of story #18405: distinguish Features and P.I. that comes from a Program in a Team

Change-Id: I27c3840289a8df63d839ce9d7cc059f2240e1596