Make possible to identify GitLab webhook requests in the metrics
The requests made to the endpoint receiving the webhook data from GitLab
instances can now be identified in the project_service_access_total
metric.

This is an interesting information to track since this endpoint will be
called to react on actions happenning outside of the Tuleap instance.

Part of story #17180 take into account gitlab commits

Change-Id: Icd3fff6f0fa5eef1f0d43ca27506fa03c99371d9