request #18487 Artifacts from deleted trackers must not be retrieved in hierarchy
If you define two trackers (A and B), with tracker A parent of tracker B (per hierarchy).
We create two artifacts (A1 and B1), and link A1 with B1.
This link will have the type _is_child because of the hierarchy.

When we delete tracker A, A1 is still displayed as the parent of B1, which must not be the case.

This kind of check was added in SQL queries that checks artifact links across services.

Change-Id: I4867e6f27e19b5609fcc3d83833d1c7dcd12851c