Display current branch/tag if git reference
If a branch is associated to the commit being referenced by (or
referencing) an artifact, then the name of the branch is displayed as
additional (primary) badge.

If no branch is associated but a tag is, then the name of the tag is
displayed as additianal (primary) badge.

Badges for tag are plain (while branches ones are outlined) in order to
be consistent with what have been done for the new git interface on a
given commit [EXAMPLE].

Since there is not much room to display additional badges for a git
commit, only the first branch or the first tag is displayed. This is the
same behavior than for current branch selector in git repository: only
the first branch or the first tag is used.

Part of story #17180 take into account gitlab commits

[EXAMLPE] https://tuleap.net/plugins/git/tuleap/tuleap/stable?a=commit&h=6a4c248f9fa0b422f766e179ba7e30c786e64158

Change-Id: I59e1019f81e0ca333e48c0b7274887c46e545078