Refactoring: centralize project access check for xrefs
Until now, each plugin is responsible to check if the reference to
display is belonging to a project the user can access. This has two
drawbacks:
* We duplicate code
* We don't enforce this check, one new plugin could omit this check

Now, everytime a cross reference is organized in a nature, the project
access is automatically checked.

No functional changes.

Part of story #17180 take into account gitlab commits

Change-Id: Iba697982c05706de59303b1e7f0e78a5199d7a7f