Refactoring: Extract representation builders
part of story #18336 add support of markdown

No functional change expected:
- in all routes that list Artifacts (GET artifacts, changesets, reports)
- in GET FRS releases route
- in Artifact webhooks (that are triggered after an update or creation
of artifact).

Instead of pushing even more responsibilities to Comment and Changeset
classes, this contribution aims to decrease them. The same way we don't
ask a DAO class to output HTML, we should not ask a data class to output
JSON representations.

This refactoring enables us to (later) inject dependencies in the
CommentRepresentationBuilder to extract references and render Markdown
in HTML.

Change-Id: Iee7da77fb3b44005ac5338fcf7312a738405af37