Do not display cross references of unreadable wiki pages
If wiki pages is not accessible to the current user, then we should not
display their references in artifact cross references field.

Note: due to direct call to db_query() in WikiPage constructor, it is
not possible to add unit test to the new dedicated object that returns a
WikiPage.

Part of story #17180 take into account gitlab commits

Change-Id: I6c418eae74a6adf29b329ab974f9fb1ee8f15f32