Display parents from another project on PV2 cards with their project background color
The color is only shown when:
 * the card has a parent from a project different than the current
   project thanks to a hierarchy defined through the scaled agile plugin
 * the project of the parent artifact has defined a background
 * the card does not have a background color

The background color of the element is derived from the proeminent color
of the background image. The way to obtain it is temporary until we have
a clue on how the design team expected us to obtain it.

Part of story #18405: distinguish Features and P.I. that comes from a Program in a Team

Change-Id: I38bdbe9029ae773a8ebb3bf80812fb9a8a098c52