Display missing icons in front of natures
Note: instead of using legacy FontAwesome 4 version, we use directly the
FontAwesome 5. This means that we cannot always reuse the icon defined
by the service, which is still based on FA4. For example for FRS,
"fa fa-files-o" becomes "far fa-copy".
Once the icons for service are migrated to FA5, we will be able to
factorize code to not duplicate icon definition.

Part of story #17180 take into account gitlab commits

Change-Id: I33c0b269dcfd9a2563d1a80f562bb796b763db3a