Do not display xref to unreachable forum message
Given an artifact is referenced by (or is referencing with `msg #123`) a
forum message,
When the user displays the artifact,
Then the subject of the message is displayed instead of the raw xref.

Given an artifact is referenced by (or is referencing with `msg #123`) a
forum message,
And the user is not allowed to access the message (private forum, or no
access to the project)
When the user displays the artifact,
Then the xref is not displayed.

Note: in order to not loose to much hair by understanding/refactoring
existing code in order to retrieve a given message, it has been moved
as-is in a dedicated object. This object is using db_query and legacy
functions, so no unit tests have been added. There should be no
functional changes on the display of a forum message.

Part of story #17180 take into account gitlab commits

Change-Id: I36a322c3df34159ce8a437d3576de259d797e2cd