Remove _is_child at tracker deletion
This is part of request #18487 Artifacts from deleted trackers must not be retrieved in hierarchy

When having two trackers (Parent A and Chil B), and two artifatcs (A1
and B1).
When linking A1 to B1, _is_child is automatically added because of the
hierarchy.

When deleting Tracker A, the link between A1 and B1 must be kept (in
case we restore the tracker), but the type _is_child must be remove
because the hierarcy between Tracker A and Tracker B is removed.

Change-Id: Id75c177827a8c6d4ea99d7dafdfa7540c2748097