Do not display xref to unreachable forum
Given an artifact is referencing (with `forum #123`) a forum,
When the user displays the artifact,
Then the name of the forum is displayed instead of the raw xref.

Given an artifact is referencing (with `forum #123`) a forum,
And the user is not allowed to access the forum (private forum, or no
access to the project, or deleted forum)
When the user displays the artifact,
Then the xref is not displayed.

Part of story #17180: take into account gitlab commits

Change-Id: I52e68d1af563eda42c8c4164954a24fd0092939a