request #18386 Unable to delete time entry in Timetracking widget
When I click on the "Delete" icon next to a time entry in Timetracking widget,
a confirmation "popover" shows up. When I confirm again, nothing happens.
The time entry is not deleted, there is no (related) javascript error in the console,
no outgoing REST request.

I expected the second confirmation to actually delete the time entry.

Change-Id: Id0efcca0f0312fc4b6f8fc66bed1aad7ec93e6b8