Display xref sparklines next to the title
With the new layout of the cross reference field in artifact, if a
reference has sparklines, then the image is displayed above the title
instead of at the left.

You can test with a reference to successfull build:
* Have the continuous integration service activated in your project
* Add a job which target a Jenkins job that contains successful build
(note its number)
* In the follow-up comments of the artifact add a cross reference to the
build: build #123 or build #my_job_name/123
* Submit and stay
=> in the follow-up comment, nothing changed, the blue bullet is still
display at the left of the xref
=> in the cross reference field (you will need to click the switch), the
reference to the build has the bullet at the left of the xref instead of
above.

Part of story #17180: take into account gitlab commits

Change-Id: I40ec14ff2dab7d0fbaa67de550b9d083977aec8d