request #18499 Containers that didn't go through install cannot do svn
Given you have a container that is running official TEE docker image
but that was not installed via the TEE docker image installation process
(migration from other image) the init_mode will not be set.

Hence when it comes to restart apache for svn usage, Tuleap won't know
how to do it and the new svn repositories will not be seen.

Even if it's a one time operation/correction I choose to set it at each
start mainly to avoid to have specific code in mytuleap migration to deal
with this situation but also to ensure that whenever the container is
restarted, regardless of what might have been done, the configuration is
restored to something consistent given that container is under supervisord
control.

Who know when it will come to haunt us 👻

Change-Id: Id3f19efe5a1fa8c24217690100b1bf7b966ef38e