Disable keyboard shortcuts when a modal is open
story #18401

When navigating with the mouse, the modal blocks much of the view and blocks the global context. I cannot click "under" the modal. Similarly, I should not be able to trigger "global" shortcuts that have an effect outside of the modal when a modal is open.

Instead of closing the modal when triggering a shortcut, we disable shortcuts when a modal is open.

Change-Id: Ia85be92401e03369d6e1c33c59154401c70cbd67