Do not display svn cross reference in private projects
When you are not member of a private project, then svn revision are not
accessible. The corresponding cross reference should not be displayed in
the artifact.

Implementation note:
Instead of adding a dedicated "SubversionOrganizer" (like it has been
done for News or Forum), we now check the project access for every cross
references, even those which don't have a dedicated "organizer". This
means that:
Given any cross reference from/to an artifact
When the user cannot access the target project
Then the cross reference is not displayed

Part of story #17180: take into account gitlab commits

Change-Id: I10e2d822d9ac7edeb23bd0686d019d9f6b3c7d7c