Remove dependency on GitPHP\Commit for CrossReferenceGitEnhancer
This is a refactoring. No functional changes are expected.

CrossReferenceGitEnhancer is using GitPHP\Commit and GitRepository only
to retrieve a CommitDetails. Instead we can pass directly the
CommitDetails.

This will help for future refactoring in order to reduce the number of
sql queries to retrieve the contributors of referenced commits.

Part of story #17180: take into account gitlab commits

Change-Id: Ia5d88033afd02afe1278c644d32d0564cbeef61b