Move relative-date outside of TLP
Part of request #18842 Split TLP into separate smaller libraries

It cannot be its own library due to dynamic import.
It is not included in TLP but only in certain BurningParrot pages. The
main problem is its polyfill dependencies: things are easier to handle
when they are not included in TLP.

Change-Id: I09e775f64d535e86728415d9ad2d2e7c35bf82b1