Introduce Roadmap widget
The widget asks for a tracker id.

It displays a "Under construction" message since it does nothing yet.

When a project is created from a template using this widget, the new
widget targets the new tracker, instead of the one in the template.
(like it is done in request #19329, request #19338, or request #19360).

The widget is hidden behind a feature flag:

```
tuleap config-set feature_flag_plugin_roadmap_display_underconstruction_widget 1
```

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: Iece0ddb1d72750be7df55013c903210e1923a39b