Move tlp-doc outside of TLP
Part of request #18842 Split TLP into separate smaller libraries

tlp-doc cannot be a lib, as it outputs files with a manifest.json and
hash in their filenames.
It should however be separate from tlp's source, because it depends on
much more than TLP: it also depends on relative-date custom element,
list-picker, codemirror, dompurify.

Change-Id: Ie16144c710d10f47ab19537f5b5615b35b6b6299