Add a `skip to main content` link as first focusable element
Part of story #18322: Quickly access a page's main content when navigating with the keyboard

When navigating with the keyboard, the first tabbed element is a "Skip to main content" link.
It targets the main content of the page, so that the user can reach it without tabbing through all the menu elements.

Change-Id: I73ff5ecfe9686f95cd24826bd01ad07e55c94d97