Extract card-fields lib
Part of story #18336 add support of markdown

No functional change expected in card fields display in Kanban and
Planning view.

Extracting this as a lib lets us remove some of the aliasing "magic"
performed both in Jest and in Webpack. Not all could be removed though
because there are still a couple of "external" dependencies (the
Artifact modal, the cumulative flow diagram).

Change-Id: Iea616b371c07abe26d9eedc1a7123dc94f22262f