Extract parser references
This is part of story #17182 reference tuleap artifacts in gitlab merge requests

Initially, the parser of reference was used by only commit, but in next patch it will be used by merge request.
To allow DRY, it is now commonly, and can be used by commit and merge request.

No function changes.
When you commit with a tuleap reference, the reference is extracted and a cross reference is created.

How to test:
  - Have gitlab instance
  - Have integrated repository
  - Commit in this repository with a tuleap reference (TULEAP-XXX where XXX is an artifact id)

Expected results:
  - A xRef is created is the artifact

Change-Id: Ic485af5573c513ff8a5e7f35ba76fe5f95151ba5